Skip Flyway provisioning if ojdbc6.jar is not present.

+++ Vagrantfile
@@ -33,9 +33,11 @@ Vagrant.configure("2") do |config|
   config.vbguest.auto_update = false
 
   $install_puppet_modules = <<SCRIPT
-  mkdir -p /etc/puppet/modules
-  puppet module list | grep -q puppetlabs-java || puppet module install puppetlabs-java
-  puppet module list | grep -q maestrodev-maven || puppet module install maestrodev-maven
+  if [ -f /home/vagrant/vagrant-ubuntu-oracle-xe/oracle-jdbc/ojdbc6.jar ]; then
+    mkdir -p /etc/puppet/modules
+    puppet module list | grep -q puppetlabs-java || puppet module install puppetlabs-java
+    puppet module list | grep -q maestrodev-maven || puppet module install maestrodev-maven
+  fi
 SCRIPT
 
   config.vm.provision "shell", inline: $install_puppet_modules
